requirements.txt
langchain
openai
python-dotenv
langchain_community
pypdf
tiktoken
faiss-cpu
huggingface-hub
sentence-transformers
groq
tqdm
streamlit